+++ packages/runtime/adapters/eventstore-adapters/eventstore-base/src/index.ts
@@ -5,7 +5,6 @@ import wrapMethod from './wrap-method'
 import wrapEventFilter from './wrap-event-filter'
 import wrapDispose from './wrap-dispose'
 import validateEventFilter from './validate-event-filter'
-import { MAINTENANCE_MODE_AUTO, MAINTENANCE_MODE_MANUAL } from './constants'
 import ConcurrentError from './concurrent-error'
 import {
   ResourceAlreadyExistError,
@@ -47,6 +46,20 @@ import {
   AdapterConfig,
   AdapterConfigSchema,
 } from './types'
+
+export {
+  threadArrayToCursor,
+  cursorToThreadArray,
+  initThreadArray,
+  emptyLoadEventsResult,
+} from './cursor-operations'
+export {
+  MAINTENANCE_MODE_AUTO,
+  MAINTENANCE_MODE_MANUAL,
+  THREAD_COUNT,
+  CURSOR_BUFFER_SIZE,
+  THREAD_COUNTER_BYTE_LENGTH,
+} from './constants'
 
 const wrappedCreateAdapter = <
   ConnectedProps extends AdapterPoolConnectedProps,
@@ -99,8 +112,6 @@ export {
   AlreadyFrozenError as EventstoreAlreadyFrozenError,
   AlreadyUnfrozenError as EventstoreAlreadyUnfrozenError,
   ReplicationAlreadyInProgress,
-  MAINTENANCE_MODE_AUTO,
-  MAINTENANCE_MODE_MANUAL,
   throwBadCursor,
   getNextCursor,
   snapshotTrigger,
@@ -129,6 +140,7 @@ export {
   EventThreadData,
   Cursor,
   EventsWithCursor,
+  EventWithCursor,
   EventFilter,
   ReplicationStatus,
   ReplicationState,