+++ packages/runtime/adapters/eventstore-adapters/eventstore-base/src/types.ts
@@ -98,6 +98,11 @@ export type EventsWithCursor = {
   events: SavedEvent[]
 }
 
+export type EventWithCursor = {
+  cursor: Cursor
+  event: SavedEvent
+}
+
 const EventFilterCommonSchema = t.intersection([
   t.type({
     limit: t.Int,
@@ -538,6 +543,11 @@ export interface AdapterFunctions<
     ConnectedProps,
     NonNullable<Adapter['resetReplication']>
   >
+
+  getCursorUntilEventTypes?: PoolMethod<
+    ConnectedProps,
+    NonNullable<Adapter['getCursorUntilEventTypes']>
+  >
 }
 
 export interface Adapter {
@@ -545,7 +555,7 @@ export interface Adapter {
   importEvents: (options?: Partial<ImportOptions>) => ImportEventsStream
   exportEvents: (options?: Partial<ExportOptions>) => ExportEventsStream
   getLatestEvent: (filter: EventFilter) => Promise<SavedEvent | null>
-  saveEvent: (event: InputEvent) => Promise<void>
+  saveEvent: (event: InputEvent) => Promise<EventWithCursor>
   init: () => Promise<void>
   drop: () => Promise<void>
   dispose: () => Promise<void>
@@ -614,4 +624,9 @@ export interface Adapter {
   setReplicationPaused?: (pause: boolean) => Promise<void>
   getReplicationState?: () => Promise<ReplicationState>
   resetReplication?: () => Promise<void>
+
+  getCursorUntilEventTypes?: (
+    cursor: Cursor,
+    untilEventTypes: Array<InputEvent['type']>
+  ) => Promise<string>
 }