+++ packages/runtime/adapters/eventstore-adapters/eventstore-lite/src/index.ts
@@ -42,6 +42,7 @@ import setReplicationIterator from './set-replication-iterator'
 import setReplicationPaused from './set-replication-paused'
 import getReplicationState from './get-replication-state'
 import resetReplication from './reset-replication'
+import getCursorUntilEventTypes from './get-cursor-until-event-types'
 
 import type { Adapter } from '@resolve-js/eventstore-base'
 import type { ConnectionDependencies, SqliteAdapterConfig } from './types'
@@ -87,6 +88,7 @@ const createSqliteAdapter = (options: SqliteAdapterConfig): Adapter => {
       setReplicationPaused,
       getReplicationState,
       resetReplication,
+      getCursorUntilEventTypes,
     },
     { sqlite, tmp, os, fs } as ConnectionDependencies,
     options