+++ packages/tools/testing-tools/src/constants.ts
@@ -1,1 +1,3 @@
 export const testContext = Symbol()
+export const ambiguousEventsTimeErrorMessage =
+  'The givenEvents function requires either all or none event timestamps to be specified.'