+++ packages/tools/testing-tools/src/flow/saga/make-test-environment.ts
@@ -136,6 +136,7 @@ export const makeTestEnvironment = (
         scheduledCommands: [],
         queries: [],
         sideEffects: [],
+        scheduleCommands: [],
       },
       // FIXME: deprecated
       'scheduleCommands',
@@ -194,12 +195,13 @@ export const makeTestEnvironment = (
       error: () => void 0,
       publish: async () => void 0,
     }
-    const eventstoreAdapter = await getEventStore(events)
 
     const errors: Error[] = []
     let executor = null
 
     try {
+      const eventstoreAdapter = await getEventStore(events)
+
       const runtime = getSagaRuntime(
         result,
         {